% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchCreate.R
\name{create_Pitch}
\alias{create_Pitch}
\title{Create a pitch plot ready for Statsbomb data}
\usage{
create_Pitch(grass_colour = "#F9F9F9", line_colour = "#8F8F8F",
  background_colour = "#F9F9F9", goal_colour = "#000000",
  goaltype = "line", middlethird = FALSE, BasicFeatures = FALSE,
  JdeP = FALSE, padding = 5)
}
\arguments{
\item{grass_colour}{A string as a colour by word or #HEX code}

\item{line_colour}{A string as a colour by word or #HEX code}

\item{background_colour}{A string as a colour by word or #HEX code}

\item{goal_colour}{A string as a colour by word or #HEX code}

\item{goaltype}{String, either "line", "box" or "barcanumbers". Default "line"}

\item{middlethird}{Boolean. Default FALSE}

\item{BasicFeatures}{Boolean. Default FALSE}

\item{JdeP}{Boolean. Default FALSE}

\item{padding}{Numeric. Default = 5}
}
\value{
A plot of a pitch
}
\description{
Creating a pitch plot using ggplot2 that is ready to plot actions on top. Numerous options allow customisation.
}
